+++ src/args/argument_list.go
@@ -27,6 +27,7 @@ type ArgumentList struct {
 	CustomMetricsConfig          string `default:"" help:"YAML configuration with one or more SQL queries to collect custom metrics"`
 	ShowVersion                  bool   `default:"false" help:"Print build information and exit"`
 	ExtraConnectionURLArgs       string `default:"" help:"Appends additional parameters to connection url. Ex. 'applicationintent=readonly&foo=bar'"`
+	EnableQueryPerformance       bool   `default:"true"`
 }
 
 // Validate validates SQL specific arguments
