+++ src/queryAnalysis/utils.go
@@ -1,9 +1,10 @@
-package queryhandler
+package queryAnalysis
 
 import (
 	_ "embed"
 	"encoding/json"
 	"fmt"
+	"regexp"
 	"strconv"
 
 	"github.com/jmoiron/sqlx"
@@ -12,14 +13,10 @@ import (
 	"github.com/newrelic/nri-mssql/src/queryAnalysis/models"
 )
 
-var _ QueryHandler = (*QueryHandlerImpl)(nil)
-
-type QueryHandlerImpl struct{}
-
-//go:embed queries.json
+//go:embed config/queries.json
 var queriesJSON []byte
 
-func (q *QueryHandlerImpl) LoadQueries() ([]models.QueryDetailsDto, error) {
+func LoadQueries() ([]models.QueryDetailsDto, error) {
 	var queries []models.QueryDetailsDto
 	if err := json.Unmarshal(queriesJSON, &queries); err != nil {
 		return nil, fmt.Errorf("failed to unmarshal queries configuration: %w", err)
@@ -42,7 +39,7 @@ func LoadQueryResponseModel(queryType string) (interface{}, error) {
 	}
 }
 
-func (q *QueryHandlerImpl) ExecuteQuery(db *sqlx.DB, queryDetailsDto models.QueryDetailsDto) ([]interface{}, error) {
+func ExecuteQuery(db *sqlx.DB, queryDetailsDto models.QueryDetailsDto) ([]interface{}, error) {
 	fmt.Println("Executing query...", queryDetailsDto.Name)
 
 	rows, err := db.Queryx(queryDetailsDto.Query)
@@ -50,11 +47,11 @@ func (q *QueryHandlerImpl) ExecuteQuery(db *sqlx.DB, queryDetailsDto models.Quer
 		return nil, fmt.Errorf("failed to execute query: %w", err)
 	}
 
-	return q.BindQueryResults(rows, queryDetailsDto)
+	return BindQueryResults(rows, queryDetailsDto)
 }
 
 // BindQueryResults binds query results to the specified data model using `sqlx`
-func (q *QueryHandlerImpl) BindQueryResults(rows *sqlx.Rows, queryDetailsDto models.QueryDetailsDto) ([]interface{}, error) {
+func BindQueryResults(rows *sqlx.Rows, queryDetailsDto models.QueryDetailsDto) ([]interface{}, error) {
 	defer rows.Close()
 
 	results := make([]interface{}, 0)
@@ -98,7 +95,7 @@ func (q *QueryHandlerImpl) BindQueryResults(rows *sqlx.Rows, queryDetailsDto mod
 }
 
 // IngestQueryMetrics processes and ingests query metrics into the New Relic entity
-func (q *QueryHandlerImpl) IngestQueryMetrics(entity *integration.Entity, results []interface{}, queryDetailsDto models.QueryDetailsDto) error {
+func IngestQueryMetrics(entity *integration.Entity, results []interface{}, queryDetailsDto models.QueryDetailsDto) error {
 	for i, result := range results {
 		// Convert the result into a map[string]interface{} for dynamic key-value access
 		var resultMap map[string]interface{}
@@ -138,4 +135,12 @@ func DetectMetricType(value string) metric.SourceType {
 	}
 
 	return metric.GAUGE
+}
+
+func AnonymizeQuery(query string) (string, error) {
+	// Regular expression to match literal values in SQL queries
+	re := regexp.MustCompile(`'[^']*'|\d+|".*?"`)
+	// Replace matched values with placeholders
+	anonymizedQuery := re.ReplaceAllString(query, "?")
+	return anonymizedQuery, nil
 }
