+++ src/queryAnalysis/validation/validation_check.go
